@@ -10,17 +10,17 @@
   <packaging>war</packaging>
 
   <name>Camunda Platform Process Application</name>
-  <description>A Process Application for [Camunda Platform](http://docs.camunda.org). [The project has been generated by the Maven archetype 'camunda-archetype-ejb-war-7.22.0']</description>
+  <description>A Process Application for [Camunda Platform](http://docs.camunda.org). [The project has been generated by the Maven archetype 'camunda-archetype-ejb-war-7.23.0']</description>
 
   <properties>
     <!--
     Adjust if you want to use Camunda Enterprise Edition (EE):
-    <camunda.version>7.22.0-ee</camunda.version>
+    <camunda.version>7.23.0-ee</camunda.version>
     Make sure you also switch to EE repository below
     -->
-    <camunda.version>7.22.0</camunda.version>
+    <camunda.version>7.23.0</camunda.version>
 
-    <maven.compiler.release>11</maven.compiler.release>
+    <maven.compiler.release>17</maven.compiler.release>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <failOnMissingWebXml>false</failOnMissingWebXml>
   </properties>
@@ -45,7 +45,7 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
-      <!-- client for Java EE application server integration, included in WAR as an alternative to write your own ProcessApplication class -->
+      <!-- client for Jakarta EE application server integration, included in WAR as an alternative to write your own ProcessApplication class -->
       <groupId>org.camunda.bpm.javaee</groupId>
       <artifactId>camunda-ejb-client-jakarta</artifactId>
     </dependency>
@@ -65,7 +65,7 @@
     <dependency>
       <groupId>org.assertj</groupId>
       <artifactId>assertj-core</artifactId>
-      <version>3.24.2</version>
+      <version>3.26.3</version>
       <scope>test</scope>
     </dependency>
 
@@ -88,11 +88,11 @@
       <scope>provided</scope>
     </dependency>
 
-    <!-- Java EE 6 APIs -->
+    <!-- Jakarta EE APIs -->
     <dependency>
-      <groupId>javax</groupId>
-      <artifactId>javaee-api</artifactId>
-      <version>6.0</version>
+      <groupId>jakarta.platform</groupId>
+      <artifactId>jakarta.jakartaee-api</artifactId>
+      <version>10.0.0</version>
       <scope>provided</scope>
     </dependency>
 
@@ -106,7 +106,7 @@
       <!-- Needed for InMemoryH2Test -->
       <groupId>com.h2database</groupId>
       <artifactId>h2</artifactId>
-      <version>2.1.214</version>
+      <version>2.3.232</version>
       <scope>test</scope>
     </dependency>
 
